pyproject.toml
[project]
name = "strands-demo"
version = "0.1.0"
description = "AWS Strands Agents demo with Streamlit frontend and AWS Cognito authentication"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Gaurav Dhamija", email = "gauravdhamija83@gmail.com" },
]
dependencies = [
    "strands-agents>=0.1.0",
    "strands-agents-tools>=0.1.0",
    "streamlit>=1.35.0",
    "boto3>=1.34.0",
    "python-dotenv>=1.0.0",
    "authlib>=1.3.2",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]
